Package structure improved from the user perspective
